requirements.txt
streamlit>=1.25.0
patronus>=0.1.0
pandas>=1.5.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
nltk>=3.8.0
scikit-learn>=1.2.0
plotly>=5.14.0
python-dotenv>=1.0.0
vaderSentiment>=3.3.2
pillow>=9.5.0
wordcloud>=1.8.2.2